backend/app/requirements.txt
psycopg2==2.9.3
gunicorn==20.1.0
coverage==6.4.4
flake8==6.0.0
django==4.2
django-extensions==3.2.0
djangorestframework==3.14.0
django-countries==7.4.2
requests==2.32.3
sib-api-v3-sdk==7.5.0
django-s3direct==2.0.3
boto3==1.24.60
admin-totals==1.0.1
django-taggit==3.0.0
factory-boy==3.2.1
Faker==15.3.4
django-apptemplates==1.5
celery==5.3.1
django-constance==2.9.1
sentry-sdk==1.13.0
django-modeltranslation==0.18.7
django-mjml==1.1
django-fullurl==1.3
django-admin-sortable2==2.1.10
django-cors-headers==3.13.0
django-versatileimagefield==2.2
django-redis==5.2.0
djangorestframework-simplejwt==5.2.2
drf-extensions==0.7.1
django-silk==5.0.3
django-debug-toolbar==3.8.1
django-json-widget==2.0.1
django-filter==24.2
drf-yasg==1.21.6
coreapi==2.3.3
pandas==2.0.3
openpyxl==3.1.2
Pillow==9.2.0
stripe==6.4.0
geos==0.2.3
pip==23.3.1
django-import-export==3.2.0
weasyprint==60.0
django-rosetta==0.9.9
django-admin-interface==0.28.2
django-colorfield==0.11.0
django-ckeditor==6.7.0
Pygments==2.17.2
opencv-python==4.9.0.80
django-admin-rangefilter==0.12.0
django-jsonform==2.22.0
